extract methods from pypescript in order to declutter. placing all utility functions in a common pypes file
